Remove 'isRead' field and filter the feeds through SQL

+++ androidApp/src/main/kotlin/com/prof18/feedflow/home/components/HomeAppBar.kt
@@ -37,7 +37,7 @@ import dev.icerock.moko.resources.compose.stringResource
 @Suppress("LongParameterList")
 @Composable
 internal fun HomeAppBar(
-    unReadCount: Int,
+    unReadCount: Long,
     showDrawerMenu: Boolean,
     isDrawerOpen: Boolean,
     onDrawerMenuClick: () -> Unit,
